+++ CHANGELOG.md
@@ -1,3 +1,7 @@
 # vue-histogram-slider changelog
 
+## v0.2.0
+- `min`, `max` props are not required
+- `type` 'single' option is completed
+
 ## v0.1.0